requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
-e git+https://github.com/digital-land/validators.git@Heroku#egg=epValidators
    # via -r requirements/requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
blinker==1.6.2
    # via flask
cchardet==2.1.7
    # via epvalidators
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via
    #   flask
    #   uvicorn
fastapi==0.95.1
    # via -r requirements/requirements.in
flask==2.3.2
    # via epvalidators
gunicorn==20.1.0
    # via -r requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httpx==0.24.0
    # via -r requirements/requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   -r requirements/requirements.in
    #   flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
numpy==1.24.3
    # via
    #   pandas
    #   shapely
pandas==2.0.1
    # via
    #   -r requirements/requirements.in
    #   epvalidators
pydantic==1.10.7
    # via fastapi
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements/requirements.in
python-magic==0.4.27
    # via epvalidators
python-multipart==0.0.6
    # via -r requirements/requirements.in
pytz==2023.3
    # via pandas
shapely==2.0.1
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
tzdata==2023.3
    # via pandas
uvicorn==0.22.0
    # via -r requirements/requirements.in
werkzeug==2.3.4
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
